% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPathoID.R
\name{assignPathowinID}
\alias{assignPathowinID}
\title{Use connection to 'activity list' to extract}
\usage{
assignPathowinID(current_PCRID, activitylisttable)
}
\arguments{
\item{current_sample}{}

\item{LUT}{}
}
\description{
Use connection to 'activity list' to extract
}
