% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{element_extracts}
\alias{element_extracts}
\title{Function to extract three strings of information from sampleID.txt}
\usage{
element_extracts(origstring)
}
\arguments{
\item{origstring}{}
}
\description{
Function to extract three strings of information from sampleID.txt
}
