% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_checks.R
\name{id_check}
\alias{id_check}
\title{Checks if ID has been observed before in sampleID table}
\usage{
id_check(current_sample, idtable)
}
\arguments{
\item{current_sample}{}

\item{idtable}{}
}
\description{
Checks if ID has been observed before in sampleID table
}
