% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_checks.R
\name{identity_check}
\alias{identity_check}
\title{Combines functions call to SNP and PatID check and}
\usage{
identity_check(current_sample, idtable)
}
\arguments{
\item{current_sample}{}

\item{idtable}{}
}
\description{
Combines functions call to SNP and PatID check and
}
