% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tables.R
\name{rowbind_tables}
\alias{rowbind_tables}
\title{Considering boolean checks for SNP and ID generates dataframe of entry sampleID and databases entries}
\usage{
rowbind_tables(current_sample, idtable)
}
\arguments{
\item{current_sample}{}

\item{idtable}{}
}
\description{
Considering boolean checks for SNP and ID generates dataframe of entry sampleID and databases entries
}
