% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{sample_info_extract}
\alias{sample_info_extract}
\title{Reads in single line sampleID.txt file, applies element_extracts and generates dataframe}
\usage{
sample_info_extract(orig)
}
\arguments{
\item{sampleIDpath}{}
}
\description{
Reads in single line sampleID.txt file, applies element_extracts and generates dataframe
}
