% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_out_result.R
\name{write_out_result}
\alias{write_out_result}
\title{Write out table to QC_SNP_PathowinPatientID.txt}
\usage{
write_out_result(df, filepath)
}
\arguments{
\item{df}{}

\item{filepath}{}
}
\description{
Write out table to QC_SNP_PathowinPatientID.txt
}
