% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_checks.R
\name{snp_check}
\alias{snp_check}
\title{Checks if SNP has been observed before and returns sampleID table entry or entries that match.}
\usage{
snp_check(sample_SNP, idtable)
}
\arguments{
\item{idtable}{}

\item{current_sample}{}
}
\description{
Checks if SNP has been observed before and returns sampleID table entry or entries that match.
}
